minor fixes

+++ src/gui/wxComposeView.cpp
@@ -350,10 +350,10 @@ wxComposeView::CreateFTCanvas(void)
       bg = READ_CONFIG(m_Profile,MP_FTEXT_BGCOLOUR);
    
    m_LayoutWindow->Clear(
-      READ_CONFIG(m_Profile,(int)MP_FTEXT_FONT),
-      READ_CONFIG(m_Profile,(int)MP_FTEXT_SIZE),
-      READ_CONFIG(m_Profile,(int)MP_FTEXT_STYLE),
-      READ_CONFIG(m_Profile,(int)MP_FTEXT_WEIGHT),
+      READ_CONFIG(m_Profile,MP_FTEXT_FONT),
+      READ_CONFIG(m_Profile,MP_FTEXT_SIZE),
+      READ_CONFIG(m_Profile,MP_FTEXT_STYLE),
+      READ_CONFIG(m_Profile,MP_FTEXT_WEIGHT),
       0,
       fg.c_str(),bg.c_str());
    m_LayoutWindow->GetLayoutList().SetEditable(true);
